Removed old navigation process
